feat(MarathonActions): delete SDK Services before deleting group
When we delete a group the services are correctly deleted as well. However, Cosmos and Marathon don't communicate with each other, so if we have SDK Services in the group these aren't correctly uninstalled.
Solution: We'll have to first uninstall the SDK Services from the group and then delete the group.

Please see the Gif below for more details.

Closes DCOS-18006